login page and root layout moved on the server side ---  routering logics moved on the user provider
