+++ app/layout.jsx
@@ -1,5 +1,4 @@
 /* eslint-disable react-hooks/exhaustive-deps */
-"use client";
 
 import { UserProvider } from "./context/user-provider";
 import "./globals.css";
