+++ app/page.jsx
@@ -1,26 +1,9 @@
 /* eslint-disable react-hooks/exhaustive-deps */
-"use client";
 
-import { useRouter } from "next/navigation";
 import Logo from "./components/Logo";
 import AuthForm from "./components/forms/AuthForm";
-import { useUser } from "./context/user-provider";
-import { useEffect } from "react";
 
 const SignIn = () => {
-  const { Profile } = useUser();
-  const router = useRouter();
-
-  useEffect(() => {
-    if (Profile) {
-      if (Profile.isAdmin) {
-        router.push("/admin");
-      } else {
-        router.push("/dashboard");
-      }
-    }
-  }, [Profile]);
-
   const inputs = [
     {
       label: "Email",
@@ -41,7 +24,7 @@ const SignIn = () => {
         title={"Welcome Student"}
         subtitle={"Please enter your credentials"}
         actionLabel={"Sign In"}
-        onRedirect={() => router.push("/sign-up")}
+        // onRedirect={() => router.push("/sign-up")}
         redirectLabel={"Don't have an account?"}
         redirectSpan={"Create"}
         inputs={inputs}
